+++ backend/core/src/main/java/com/scoinone/core/dto/response/order/GetSellOrderListResponseDto.java
@@ -2,6 +2,8 @@ package com.scoinone.core.dto.response.order;
 
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -15,8 +17,8 @@ public class GetSellOrderListResponseDto {
         private String orderId;
         private String sellerId;
         private String virtualAssetId;
-        private Double quantity;
-        private Double price;
+        private BigDecimal quantity;
+        private BigDecimal price;
         private String status;
         private String tradeTime;
         private String createdAt;
